changing into to float for task 10

+++ MySQL_Advanced/10-div.sql
@@ -1,7 +1,7 @@
 -- A safe division function that returns 0 if the second argument is 0
 
 DELIMITER //
-CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS INT
+CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS FLOAT
 BEGIN
     IF b = 0 THEN
         RETURN 0;
